spelling, naming, warnings

+++ util/src/fragments.cpp
@@ -11,7 +11,7 @@
 #include <cstring>
 
 void copyRange(uint8_t* destination, FragmentList src, size_t skip, size_t size) {
-	int fragmentIndex = 0;
+	size_t fragmentIndex = 0;
 
 	// Find which fragment to start - skip any full fragments smaller than `skip` parameter
 	while (skip > src.fragments[fragmentIndex].size && fragmentIndex <= src.count) {
@@ -19,7 +19,7 @@ void copyRange(uint8_t* destination, FragmentList src, size_t skip, size_t size)
 		fragmentIndex++;
 	}
 
-	int destinationIndex = 0;
+	size_t destinationIndex = 0;
 
 	while (size > 0) {
 		if (fragmentIndex >= src.count) {
